backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
prisma
groq
openai
httpx
duckduckgo-search
